Splitted NavigatorDefaults and NavigatorOptions.
